Factor out testing logic to enable sharing with the async code.
